# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Magenta ASOAR API
  description: |
    Agentic System Orchestration Automation and Response — REST API.

    Provides endpoints for agent management, mission orchestration,
    log ingestion, data mesh queries, memory operations, and system health.
  version: 0.1.0
  contact:
    name: Magenta SOC Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://magenta-api.magenta-soa.svc.cluster.local
    description: Kubernetes cluster
...
paths:
...
  /api/v1/mesh/query:
    post:
      tags: [Data Mesh]
      summary: Hybrid search across data products
      operationId: meshQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                products:
                  type: array
                  items:
                    type: string
                filters:
                  type: object
                top_k:
                  type: integer
                  default: 10
                hybrid:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Search results with federation metadata
  /api/v1/mesh/ingest:
    post:
      tags: [Data Mesh]
      summary: Ingest documents into a data product
      operationId: meshIngest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product, documents]
              properties:
                product:
                  type: string
                documents:
                  type: array
                  items:
                    type: object
                batch_size:
                  type: integer
                  default: 32
      responses:
        "200":
          description: Ingest result
  /api/v1/mesh/memory/write-episode:
    post:
      tags: [Memory]
      summary: Write episodic memory
      operationId: writeEpisode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_role, mission_id, turn_number, text]
              properties:
                agent_role:
                  type: string
                mission_id:
                  type: string
                turn_number:
                  type: integer
                text:
                  type: string
                correlation_id:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Episode written
  /api/v1/mesh/memory/search-episodes:
    post:
      tags: [Memory]
      summary: Search episodic memory
      operationId: searchEpisodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                agent_role:
                  type: string
                mission_id:
                  type: string
                top_k:
                  type: integer
                  default: 5
      responses:
        "200":
          description: Episodic memory results
  /api/v1/mesh/memory/write-semantic:
    post:
      tags: [Memory]
      summary: Write semantic memory
      operationId: writeSemantic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                product:
                  type: string
                  default: agent.memory.semantic
                source:
                  type: string
                  default: agent
                tags:
                  type: array
                  items:
                    type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Semantic memory written
  /api/v1/mesh/memory/search-semantic:
    post:
      tags: [Memory]
      summary: Search semantic memory
      operationId: searchSemantic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                product:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                top_k:
                  type: integer
                  default: 5
      responses:
        "200":
          description: Semantic memory results
  /api/v1/mesh/memory/write-procedure:
    post:
      tags: [Memory]
      summary: Write procedural memory
      operationId: writeProcedure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool_name, text]
              properties:
                tool_name:
                  type: string
                text:
                  type: string
                parameters:
                  type: object
                mission_id:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Procedural memory written
  /api/v1/mesh/memory/search-procedures:
    post:
      tags: [Memory]
      summary: Search procedural memory
      operationId: searchProcedures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                tool_name:
                  type: string
                top_k:
                  type: integer
                  default: 5
      responses:
        "200":
          description: Procedural memory results
components:
  securitySchemes:
    hmacAuth:
      type: apiKey
      in: header
      name: X-Magenta-Signature
      description: HMAC-SHA256 signature
...